ui/requirements.txt
streamlit
pyyaml
streamlit-autorefresh
